dius: make use of binary IPC over unix sockets
This change primarily consists of marking tons of operation
states immovable, as a side of effect of me trying to figure
out why the test code worked on Clang but not GCC.

It turns out the culprit was the `MessageWithReply` concept,
which returned true on GCC but false with Clang for the following
nested type:

```cpp
template<typename T>
concept MessageWithReply = requires{ typename T::Reply; };

struct Message {
    struct Reply {};
};
```

On GCC, Message::Reply itself was considered to have a Reply,
for whatever reason (I don't know which behavior is correct).
The fix is to also ensure that the Reply is not itself.